Cleanups (#65)
