+++ pytest.ini
@@ -1,4 +1,4 @@
 [pytest]
 filterwarnings =
-    ignore:The zyte_api\.aio\.client module is deprecated:DeprecationWarning
+    ignore:The zyte_api\.aio module is deprecated:DeprecationWarning
