+++ setup.cfg
@@ -10,3 +10,5 @@ per-file-ignores =
     # imports are there to expose submodule functions so they can be imported
     # directly from that module
     zyte_api/__init__.py:F401
+    zyte_api/aio/errors.py:F401
+    zyte_api/aio/retry.py:F401
