+++ zyte_api/__init__.py
@@ -3,4 +3,6 @@ Python client libraries and command line utilities for Zyte API
 """
 
 from ._async import AsyncZyteAPI
+from ._errors import RequestError
+from ._retry import RetryFactory, zyte_api_retrying
 from ._utils import create_session
